// VIS_Application/app/Model/Masters/CompanyRelated/CompanyMaster.ts
export interface ICompanyMaster {

    Id: number;
    CompanyName: string;
    City: string;
    ZipCode: string;
    State: string;
    Country: string;
    CountryName: string;
    ContactNo: string;
    E_mail: string;
    Fax: string;
    CompanyURL: string;
    Address: string;
    CompanyLogo: string;
    CompanyLogoUpload : string;
    AccountHead: string;
    Designation: string;
    Signature: string;
    SignatureUpload: string;
    CompanyShortCode: string;
    Address2: string;
    StartSeries: string;
    FinancialYear: number;
    AccountHeadOther: string;
    DesignationOther: string;
    SignatureOther: string;
    SignatureOtherUpload: string;
    CreatedOn: Date;
    CreatedBy: string;
    UpdatedOn: Date;
    UpdatedBy: string
    IsActive: boolean;

    PositionName: string;
    Employee_Name: string;
    FinancialYeardll: string;
    
}